+++ src/lib.rs
@@ -172,6 +172,11 @@ pub fn get_picture_from_resources(image_name: &str) -> gtk::Picture {
             let loader = gdk_pixbuf::PixbufLoader::new();
 
             if loader.write(&image_bytes).is_ok() {
+                match loader.close() {
+                    Ok(_) => {}
+                    Err(error) => eprintln!("\t- ERROR problem with loader:\n\t{:?}", error),
+                };
+
                 let texture = gtk::gdk::Texture::from_bytes(&image_bytes)
                     .map_err(|err| format!("Failed to create texture: {}", err))
                     .unwrap();